refactor: updates css implementation
