+++ src/assets/styles/components.scss
@@ -282,3 +282,6 @@ iframe {
 @import 'src/components/LoginForm';
 @import 'src/components/BackButton';
 @import 'src/components/Dropzone';
+
+@import 'src/components/Modal';
+@import 'src/components/Storage';
